+++ lib/perfect_shape/math.rb
@@ -15,6 +15,28 @@ module PerfectShape
       def degrees_to_radians(degrees)
         (Math::PI/180)*degrees
       end
+        
+      # Normalizes the specified angle into the range -180 to 180.
+      def normalize_degrees(angle)
+        if angle > 180.0
+          if angle <= (180.0 + 360.0)
+            angle = angle - 360.0
+          else
+            angle = Math.ieee_remainder(angle, 360.0)
+            # IEEEremainder can return -180 here for some input values...
+            angle = 180.0 if angle == -180.0
+          end
+        elsif angle <= -180.0
+          if angle > (-180.0 - 360.0)
+            angle = angle + 360.0
+          else
+            angle = Math.ieee_remainder(angle, 360.0)
+            # IEEEremainder can return -180 here for some input values...
+            angle = 180.0 if angle == -180.0
+          end
+        end
+        angle
+      end
 
       # Computes the remainder operation on two arguments as prescribed by the IEEE 754 standard.
       #