+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - `PerfectShape::Arc#center_x` / `PerfectShape::Arc#center_y` / `PerfectShape::Arc#radius_x` / `PerfectShape::Arc#radius_y`
 - `PerfectShape::Rectangle#center_x` / `PerfectShape::Arc#center_y`
+- `PerfectShape::Square#center_x` / `PerfectShape::Square#center_y`
+- `PerfectShape::Arc#initialize(center_x: , center_y: , radius_x: , radius_y: )`
+
 
 ## 0.0.3